Push dependency sources to neovim/deps. #57
Added a new build that pushes the sources of the built-in dependencies
to neovim/deps.

This is useful for the Ubuntu PPA, where the sources cannot be
downloaded during the build (no internet connections allowed) and thus
must already be present beforehand. Currently, this is done by manually
updating a Launchpad repository every time one of Neovim's builtin
dependencies is updated.