with_token: Use $token to perform replacement of '%token%' literal (#108)
The ${var//pattern/repl} form of parameter expansion interprets a
leading % as anchoring pattern at the end of $var.  This broke the
github-api functions because they need to replace "%token%" inside
"%token%:x-oauth-basic", which doesn't match at the end.

Using a variable as the pattern avoids interpreting % as an anchor,
instead treating the text as a literal string.