with_token: Combine stdout and stderr, rather than duping stdout
1062e2a09d70ff6d49cdbcb411f4801a6a9fbb51 was supposed to "Redirect
stderr", but instead it duplicated stdout to stderr.  This breaks any
consumer of output from something with_token calls because there is
nothing in stdout anymore.

Since the intent was to make sure GH_TOKEN is scrubbed from the output,
combine stderr with stdout, which is subsequently scrubbed by the
following sed command.