deps-src: Keep <proj>-stamp directory, but remove post-patch stamp files
The main steps of the ExternalProject build are, in order: mkdir,
download, update/patch, configure, build, install, done.  Since the
source directories are already patched, builds will fail if they try to
re-apply the patches.

We already run the (dist)clean targets for each project, so it should be
safe to run everything after the patch step of the ExternalProject
machinery.