ci: remove clang report page
The warnings found by clang report is already preventatively checked in
the neovim repository:
https://github.com/neovim/neovim/blob/7fa24948a936a95519f0c8c496402488b6508c14/.github/workflows/test.yml#L82-L95.
This means that new warnings will not be passed in and the need for the
report page is greatly reduced.

Furthermore, some of the warnings such as core.NullDereference seem
susceptible to false positives which are hard to identify and should be
ignored for the time being. `clang-analyzer` gives better tools compared
to `scan-build` to selectively ignore these warnings.

Lastly, this will reduce the `reports` workflow from around 9 minutes to
1 minute.